+++ src/test/groovy/ru/vyarus/dropwizard/guice/AbstractTest.groovy
@@ -1,16 +1,16 @@
 package ru.vyarus.dropwizard.guice
 
 import com.codahale.metrics.health.HealthCheckRegistry
-import com.google.common.reflect.Reflection
 import io.dropwizard.jersey.setup.JerseyEnvironment
 import io.dropwizard.jetty.MutableServletContextHandler
 import io.dropwizard.jetty.setup.ServletEnvironment
 import io.dropwizard.lifecycle.setup.LifecycleEnvironment
 import io.dropwizard.setup.AdminEnvironment
-import io.dropwizard.setup.Environment;
+import io.dropwizard.setup.Environment
 import spock.lang.Specification
 
 import javax.servlet.FilterRegistration
+import javax.servlet.ServletRegistration
 
 /**
  * Base class for tests.
@@ -35,6 +35,8 @@ abstract class AbstractTest extends Specification {
         environment.servlets().addFilter(*_) >> Mock(FilterRegistration.Dynamic)
         environment.getApplicationContext() >> Mock(MutableServletContextHandler)
         environment.admin() >> Mock(AdminEnvironment)
+        environment.admin().addFilter(*_) >> Mock(FilterRegistration.Dynamic)
+        environment.admin().addServlet(*_) >> Mock(ServletRegistration.Dynamic)
         environment.lifecycle() >> Mock(LifecycleEnvironment)
         environment.healthChecks() >> Mock(HealthCheckRegistry)
         return environment